backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.3.4
pydantic==2.5.3
python-multipart==0.0.6
httpx>=0.24.0,<0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
